(PC-29416) feat(ReactionChoiceValidation): add reactions buttons to Offer
